% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metacell_subpopulation.R
\name{filter_metacells_to_subpopulation}
\alias{filter_metacells_to_subpopulation}
\title{Filter metacell matrices to a donor subpopulation}
\usage{
filter_metacells_to_subpopulation(inDir, outDir, donorListFile)
}
\arguments{
\item{inDir}{Character scalar. Input directory containing `*metacells.txt.gz`
files.}

\item{outDir}{Character scalar. Output directory to write filtered files. Will
be created if it does not exist.}

\item{donorListFile}{Character scalar. Path to a tab-delimited file whose
first column contains donor IDs matching donor column names in the metacell
matrices. The file may have no header.}
}
\value{
Invisibly returns a character vector of output file paths written.
}
\description{
Read one or more metacell matrices from `inDir` and write filtered versions to
`outDir`, keeping only the `gene_symbol` column and donor columns listed in
`donorListFile` that are present in each file.
}
\details{
Input files are matched by the pattern `metacells\\.txt\\.gz$`. Output is
written compressed (gz) to `outDir` using the input basenames.
}
