% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-table-nse.R
\name{nse-imports}
\alias{nse-imports}
\title{data.table / ggplot2 NSE imports}
\description{
Internal roxygen block used to register imports needed for NSE constructs and
to satisfy R CMD CHECK.
}
\keyword{internal}
