% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bican.mccarroll.eqtl-package.R
\docType{package}
\name{bican.mccarroll.eqtl-package}
\alias{bican.mccarroll.eqtl}
\alias{bican.mccarroll.eqtl-package}
\title{bican.mccarroll.differentialexpression}
\description{
Tools for comparing eQTL analysis
}
\author{
\strong{Maintainer}: James Nemesh \email{nemesh@broadinstitute.org}

}
\keyword{internal}
